refactor: move session related logic into it's own module
