+++ src/lib.rs
@@ -1,32 +1,15 @@
+use session::{Aggregator, Label, Session, SessionFile};
+
 use crate::date_time::DateTime;
-use std::{
-    collections::HashSet,
-    env,
-    error::Error,
-    fs, io,
-    path::{Path, PathBuf},
-    process::Command,
-    str::FromStr,
-};
+use std::{env, error::Error, fs, io, path::PathBuf, process::Command};
 
 mod date_time;
+mod session;
 #[cfg(test)]
 mod test_utils;
 
 const CONFIG_PATH: &str = "~/.timetracker.toml";
 const CONFIG_SESSIONS_PATH: &str = "sessions_path";
-
-const SESSION_HEADING_PREFIX: &str = "# ";
-const SESSION_TITLE: &str = "Session";
-const MARKS_HEADING: &str = "## Marks";
-const MARK_HEADING_PREFIX: &str = "### ";
-const LABEL_PREFIX: &str = "- ";
-const LABEL_END: &str = "- end";
-const LABEL_SKIP: &str = "- skip";
-const LABEL_TAG: &str = "- tag";
-const LABEL_TAG_SURROUND: &str = "`";
-
-const COMMAND_VIEW_MARK_CONTENTS_SEPARATOR: &str = "---------------------------------";
 
 #[derive(PartialEq, Debug)]
 pub struct Config {
@@ -193,467 +176,6 @@ impl Action {
             name => return Err(format!("unrecognized command `{name}`"))?,
         };
         Ok(out)
-    }
-}
-
-struct Aggregator {
-    sessions: Vec<Session>,
-}
-
-impl Aggregator {
-    fn build(config: &Config) -> Result<Aggregator, Box<dyn Error>> {
-        let sessions = read_sessions_dir(&config)?
-            .iter()
-            .map(|v| -> Result<Session, Box<dyn Error>> {
-                let contents = fs::read_to_string(&v)?;
-                let file = SessionFile::build(&v, &contents)?;
-                let session = Session::from_file(&file)?;
-                Ok(session)
-            })
-            .collect::<Result<Vec<_>, _>>()?;
-        if sessions.is_empty() {
-            return Err("session directory is empty")?;
-        }
-        let aggregator = Aggregator { sessions };
-        Ok(aggregator)
-    }
-
-    /// Week - Session started in the previous week that ends in the current week is still counted to the
-    /// previous week.
-    fn view(&self) -> String {
-        let session = self
-            .sessions
-            .last()
-            .expect("must always have at least one session");
-        assert!(session.marks.len() > 0);
-        assert!(
-            !(session.marks.len() == 1 && session.marks.last().unwrap().has_label(&Label::End))
-        );
-
-        let start = DateTime::new(&session.start()).to_formatted_pretty();
-        let start_of_week = DateTime::get_start_of_week(&session.start());
-        let week_time = self
-            .sessions
-            .iter()
-            .filter(|v| v.start().timestamp_millis() - start_of_week.timestamp_millis() >= 0)
-            .fold(0, |acc, val| acc + val.get_time());
-        let week_time = DateTime::get_time_hr_from_milli(week_time);
-        let session_time = DateTime::get_time_hr_from_milli(session.get_time());
-        let mark_last = session
-            .marks
-            .last()
-            .expect("session must have at least one mark");
-        let mark_last_time = if session.is_active() {
-            let timestamp_now = DateTime::now().date.timestamp_millis();
-            let timestamp_mark = mark_last.date.timestamp_millis();
-            let timestamp = timestamp_now - timestamp_mark;
-            assert!(timestamp >= 0);
-            &DateTime::get_time_hr_from_milli(timestamp.try_into().unwrap())
-        } else {
-            "0"
-        };
-        let mark_last_contents = mark_last.to_string();
-
-        let mut str = String::new();
-        if !session.is_active() {
-            str += "No active session, last session:\n";
-        }
-        str += &format!(
-            "\
-            Start: {start}\n\
-            Week: {week_time}\n\
-            Time: {session_time}\n\
-            Mark: {mark_last_time}\n\
-            {COMMAND_VIEW_MARK_CONTENTS_SEPARATOR}\n\
-            {mark_last_contents}\
-            "
-        );
-        str
-    }
-}
-
-#[derive(PartialEq, Debug)]
-struct SessionFile {
-    path: PathBuf,
-    contents: String,
-}
-
-impl SessionFile {
-    // TODO: add checks for path
-    fn build(path: &PathBuf, contents: &str) -> Result<SessionFile, &'static str> {
-        let contents = contents.trim();
-        if !contents.starts_with(SESSION_HEADING_PREFIX) || !contents.contains(MARKS_HEADING) {
-            return Err("couldn't parse session file");
-        }
-        Ok(SessionFile {
-            path: path.clone(),
-            contents: contents.to_string(),
-        })
-    }
-
-    fn get_heading_with_contents(heading: &str, contents: &str) -> String {
-        let heading_level = SessionFile::get_heading_level(&heading);
-        let mut is_within = false;
-        let mut text = String::new();
-        for line in contents.lines() {
-            if line.starts_with("#") && SessionFile::get_heading_level(&line) <= heading_level {
-                is_within = false;
-            }
-            if line.starts_with(&heading) {
-                is_within = true;
-            }
-            if is_within {
-                text += &line.trim();
-                text += "\n";
-            }
-        }
-
-        text.trim().to_string()
-    }
-
-    fn get_heading_level(heading: &str) -> u8 {
-        let mut level = 0u8;
-        for char in heading.trim().chars() {
-            if char != '#' {
-                break;
-            }
-            level += 1;
-        }
-        level
-    }
-}
-
-#[derive(PartialEq, Debug, Clone)]
-struct Session {
-    path: PathBuf,
-    marks: Vec<Mark>,
-}
-
-impl Session {
-    fn new(config: &Config, dt: &DateTime) -> Session {
-        let mark = Mark::new(&dt.date);
-        Session {
-            path: Path::join(&config.sessions_path, format!("{}.md", dt.to_formatted())),
-            marks: vec![mark],
-        }
-    }
-
-    fn get_last(config: &Config) -> Result<Option<Session>, Box<dyn Error>> {
-        let dir = read_sessions_dir(&config)?;
-        if dir.len() == 0 {
-            return Ok(None);
-        }
-        let path = &dir[dir.len() - 1];
-        let contents = fs::read_to_string(&path)?;
-        let file = SessionFile::build(&path, &contents)?;
-        let session = Session::from_file(&file)?;
-        Ok(Some(session))
-    }
-
-    fn start(&self) -> chrono::DateTime<chrono::Local> {
-        self.marks
-            .first()
-            .expect("session must have at least one mark")
-            .date
-    }
-
-    #[allow(dead_code)]
-    fn end(&self) -> chrono::DateTime<chrono::Local> {
-        self.marks
-            .last()
-            .expect("session must have at least one mark")
-            .date
-    }
-
-    fn is_active(&self) -> bool {
-        !self
-            .marks
-            .last()
-            .expect("must always have at least one mark")
-            .has_label(&Label::End)
-    }
-
-    fn get_time(&self) -> u64 {
-        let mut acc = 0;
-        let mut mark_preceding = &self.marks[0];
-        for mark in self.marks.iter().skip(1) {
-            if !mark_preceding.has_label(&Label::Skip) {
-                acc += DateTime::get_time(&mark_preceding.date, &mark.date);
-            }
-            mark_preceding = &mark;
-        }
-        if self.is_active() && !mark_preceding.has_label(&Label::Skip) {
-            acc += DateTime::get_time(&mark_preceding.date, &DateTime::now().date);
-        }
-        acc
-    }
-
-    fn stop(&mut self, dt: &DateTime) -> Result<(), &'static str> {
-        if !self.is_active() {
-            return Err("session already ended");
-        }
-        let mut mark = Mark::new(&dt.date);
-        mark.add_label(&Label::End);
-        self.marks.push(mark);
-        Ok(())
-    }
-
-    fn mark(&mut self, dt: &DateTime) -> Result<(), &'static str> {
-        if !self.is_active() {
-            return Err("can't mark, session has already ended");
-        }
-        let mark = Mark::new(&dt.date);
-        self.marks.push(mark);
-        Ok(())
-    }
-
-    fn remark(&mut self, dt: &DateTime) {
-        let mark = self
-            .marks
-            .last_mut()
-            .expect("session must always have at least one mark");
-        mark.date = dt.date;
-    }
-
-    fn label(&mut self, label: &Label) -> bool {
-        self.marks
-            .last_mut()
-            .expect("session must always have at least one mark")
-            .add_label(&label)
-    }
-
-    fn unlabel(&mut self, label: &Label) -> bool {
-        self.marks
-            .last_mut()
-            .expect("session must always have at least one mark")
-            .remove_label(&label)
-    }
-
-    /// Returns error if the content of the current mark is not empty.
-    fn write(&mut self, text: &str) -> Result<(), ()> {
-        let mark = self
-            .marks
-            .last_mut()
-            .expect("session must always have at least one mark");
-        if !mark.contents.is_empty() {
-            return Err(());
-        }
-        mark.write(&text);
-        Ok(())
-    }
-
-    fn save(&self) -> Result<(), Box<dyn Error>> {
-        let file = self.to_file()?;
-        fs::write(&file.path, &file.contents).map_err(|e| format!("coudln't save session: {e}"))?;
-        Ok(())
-    }
-
-    // TODO: make this and all other from/to methods idiomatic using traits
-    fn from_file(file: &SessionFile) -> Result<Session, Box<dyn Error>> {
-        let mut marks: Vec<Mark> = Vec::new();
-        let marks_contents = SessionFile::get_heading_with_contents(MARKS_HEADING, &file.contents);
-        for line in marks_contents.lines() {
-            if line.starts_with(MARK_HEADING_PREFIX) {
-                let contents = SessionFile::get_heading_with_contents(&line, &marks_contents);
-                let mark = Mark::from_string(&contents)?;
-                marks.push(mark);
-            }
-        }
-        if marks.is_empty() {
-            return Err("there must be at least one mark for a session to be valid")?;
-        }
-
-        Ok(Session {
-            path: file.path.clone(),
-            marks,
-        })
-    }
-
-    fn to_file(&self) -> Result<SessionFile, &'static str> {
-        let mut contents = format!(
-            "\
-            {SESSION_HEADING_PREFIX}{SESSION_TITLE}\n\
-            \n\
-            {MARKS_HEADING}\n\
-            \n\
-            "
-        );
-        for mark in &self.marks {
-            contents += &mark.to_string();
-            contents += "\n\n";
-        }
-
-        let file = SessionFile::build(&self.path, &contents)?;
-        Ok(file)
-    }
-}
-
-#[derive(PartialEq, Debug, Clone)]
-struct Mark {
-    date: chrono::DateTime<chrono::Local>,
-    labels: HashSet<Label>,
-    contents: String,
-}
-
-impl Mark {
-    fn new(date: &chrono::DateTime<chrono::Local>) -> Mark {
-        Mark {
-            date: date.clone(),
-            labels: HashSet::new(),
-            contents: String::new(),
-        }
-    }
-
-    // TODO: remove all label manipulation methods and use the labels set directly?
-    fn add_label(&mut self, label: &Label) -> bool {
-        self.labels.insert(label.clone())
-    }
-
-    fn remove_label(&mut self, label: &Label) -> bool {
-        self.labels.remove(&label)
-    }
-
-    fn has_label(&self, label: &Label) -> bool {
-        self.labels.contains(&label)
-    }
-
-    /// Overwrites the content of this mark.
-    fn write(&mut self, text: &str) {
-        self.contents = text.to_owned();
-    }
-
-    fn erase(&mut self) {
-        self.contents = String::new();
-    }
-
-    fn from_string(contents: &str) -> Result<Mark, Box<dyn Error>> {
-        let contents = contents.trim();
-        let date = contents
-            .lines()
-            .next()
-            .map(|val| chrono::DateTime::from_str(&val[MARK_HEADING_PREFIX.len()..val.len()]))
-            .ok_or("couldn't parse mark heading")??;
-        let mut contents_without_heading = contents
-            .lines()
-            .skip(1)
-            .fold(String::new(), |acc, val| acc + "\n" + val)
-            .trim()
-            .to_owned();
-        let mut labels: HashSet<Label> = HashSet::new();
-        if contents_without_heading.starts_with(LABEL_PREFIX) {
-            for line in contents_without_heading.lines() {
-                if !line.starts_with(LABEL_PREFIX) {
-                    break;
-                }
-                labels.insert(Label::from_string(&line)?);
-            }
-            contents_without_heading = contents_without_heading
-                .lines()
-                .skip(labels.len())
-                .collect::<Vec<&str>>()
-                .join("\n")
-                .trim()
-                .to_owned();
-        }
-        Ok(Mark {
-            date,
-            labels,
-            contents: contents_without_heading,
-        })
-    }
-
-    fn to_string(&self) -> String {
-        let mut contents = format!(
-            "{MARK_HEADING_PREFIX}{}",
-            DateTime::new(&self.date).to_formatted_pretty()
-        );
-        if !self.labels.is_empty() {
-            contents += "\n";
-            contents += &self
-                .labels
-                .iter()
-                .fold(String::new(), |acc, val| acc + "\n" + &val.to_string());
-        }
-        let trimmed = self.contents.trim();
-        if !trimmed.is_empty() {
-            contents += "\n\n";
-            contents += &trimmed;
-        }
-        contents
-    }
-}
-
-#[derive(Eq, Hash, PartialEq, Debug, Clone)]
-enum Label {
-    End,
-    Skip,
-    Tag { text: String },
-}
-
-impl Label {
-    fn from_args(args: &[String]) -> Result<Label, String> {
-        if args.is_empty() {
-            return Err("not enough arguments")?;
-        }
-
-        let name = LABEL_PREFIX.to_owned() + args[0].trim();
-        if name == LABEL_END {
-            if args.len() > 1 {
-                return Err("too many arguments")?;
-            }
-            return Ok(Label::End);
-        } else if name == LABEL_SKIP {
-            if args.len() > 1 {
-                return Err("too many arguments")?;
-            }
-            return Ok(Label::Skip);
-        } else if name.starts_with(LABEL_TAG) {
-            if args.len() < 2 {
-                return Err("not enough arguments")?;
-            } else if args.len() > 2 {
-                return Err("too many arguments")?;
-            }
-            let tag = Label::Tag {
-                text: args[1].clone(),
-            };
-            return Ok(tag);
-        } else {
-            return Err(format!("couldn't parse label from args '{:?}'", &args));
-        }
-    }
-
-    fn from_string(text: &str) -> Result<Label, String> {
-        let text = text.trim();
-        if text == LABEL_END {
-            return Ok(Label::End);
-        } else if text == LABEL_SKIP {
-            return Ok(Label::Skip);
-        } else if text.starts_with(&format!("{LABEL_TAG} {LABEL_TAG_SURROUND}"))
-            && text.ends_with(LABEL_TAG_SURROUND)
-        {
-            let start = format!("{LABEL_TAG} {LABEL_TAG_SURROUND}").len();
-            let end = text.len() - LABEL_TAG_SURROUND.len();
-            let tag_text = text[start..end].trim().to_owned();
-            if tag_text.is_empty() {
-                return Err("label tag cannot be empty")?;
-            }
-            let tag = Label::Tag { text: tag_text };
-            return Ok(tag);
-        } else {
-            return Err(format!("couldn't parse label from string '{}'", text));
-        }
-    }
-
-    fn to_string(&self) -> String {
-        match self {
-            Label::End => LABEL_END.to_owned(),
-            Label::Skip => LABEL_SKIP.to_owned(),
-            Label::Tag { text } => {
-                // There shouldn't be a way to store empty string in Label::Tag.
-                assert!(!text.is_empty());
-                format!("{LABEL_TAG} {LABEL_TAG_SURROUND}{text}{LABEL_TAG_SURROUND}")
-            }
-        }
     }
 }
 
@@ -873,42 +395,6 @@ fn get_git_branch_name() -> Result<String, Box<dyn Error>> {
 mod tests {
     use super::*;
 
-    use crate::test_utils;
-    use chrono::Timelike;
-
-    fn get_template(date: &str) -> String {
-        format!(
-            "\
-            {SESSION_HEADING_PREFIX}{SESSION_TITLE}\n\
-            \n\
-            {MARKS_HEADING}\n\
-            \n\
-            {MARK_HEADING_PREFIX}{date}\
-            "
-        )
-    }
-
-    fn date_plus_secs(
-        date: chrono::DateTime<chrono::Local>,
-        secs: i64,
-    ) -> chrono::DateTime<chrono::Local> {
-        chrono::DateTime::from_timestamp_millis(date.timestamp_millis() + secs * 1000)
-            .unwrap()
-            .into()
-    }
-
-    fn minutes(count: i64) -> i64 {
-        count * 60
-    }
-
-    fn hours(count: i64) -> i64 {
-        count * 60 * 60
-    }
-
-    fn days(count: i64) -> i64 {
-        count * 24 * 60 * 60
-    }
-
     #[test]
     fn config_from_args_works() {
         let args = &[String::from("time_tracker"), String::from("start")];
@@ -1040,707 +526,5 @@ mod tests {
         assert!(Action::build("hello", &[]).is_err());
 
         Ok(())
-    }
-
-    #[test]
-    fn aggregator_view_works() {
-        let mark_start = Mark::new(&date_plus_secs(test_utils::date_default(), hours(-2)));
-        let mark_end = Mark::new(&date_plus_secs(test_utils::date_default(), minutes(-30)));
-        let mut session_third = Session {
-            path: PathBuf::from("sessions"),
-            marks: vec![mark_start, mark_end.clone()],
-        };
-        let start = DateTime::new(&session_third.start()).to_formatted_pretty();
-
-        // Gets ignored because it's not in the current week.
-        let mut session_first = Session {
-            path: session_third.path.clone(),
-            marks: vec![
-                Mark::new(&date_plus_secs(test_utils::date_default(), days(-12))),
-                Mark::new(&date_plus_secs(test_utils::date_default(), days(-9))),
-            ],
-        };
-        session_first
-            .marks
-            .last_mut()
-            .unwrap()
-            .add_label(&Label::End);
-
-        let mut session_second = Session {
-            path: session_third.path.clone(),
-            marks: vec![
-                Mark::new(&date_plus_secs(test_utils::date_default(), days(-2))),
-                Mark::new(&date_plus_secs(test_utils::date_default(), days(-1))),
-            ],
-        };
-        session_second
-            .marks
-            .last_mut()
-            .unwrap()
-            .add_label(&Label::End);
-
-        let aggregator = Aggregator {
-            sessions: vec![
-                session_first.clone(),
-                session_second.clone(),
-                session_third.clone(),
-            ],
-        };
-
-        // Goes up to current time.
-        let output = aggregator.view();
-        let lines: Vec<&str> = output.lines().collect();
-        assert_eq!(lines.len(), 6);
-        assert_eq!(lines[0], format!("Start: {start}"));
-        // Not explicitly checking if it included the current time in the calculation, just
-        // excluding the possibility that it calculated only up to the last mark.
-        assert_ne!(lines[1], "Week: 25h 30m 0s");
-        assert_ne!(lines[2], "Time: 1h 30m 0s");
-        assert_ne!(lines[3], "Mark: 0");
-        assert_eq!(lines[4], COMMAND_VIEW_MARK_CONTENTS_SEPARATOR);
-        assert_eq!(lines[5], mark_end.to_string());
-
-        session_third.marks.pop();
-        session_third.stop(&DateTime::now()).unwrap();
-        let mark = &mut session_third.marks[1];
-        mark.date = mark_end.date;
-        let mark_end = session_third.marks.last().unwrap();
-        let aggregator = Aggregator {
-            sessions: vec![
-                session_first.clone(),
-                session_second.clone(),
-                session_third.clone(),
-            ],
-        };
-        assert_eq!(
-            aggregator.view(),
-            format!(
-                "\
-                No active session, last session:\n\
-                Start: {start}\n\
-                Week: 25h 30m 0s\n\
-                Time: 1h 30m 0s\n\
-                Mark: 0\n\
-                {COMMAND_VIEW_MARK_CONTENTS_SEPARATOR}\n\
-                {}\
-                ",
-                mark_end.to_string()
-            )
-        );
-    }
-
-    #[test]
-    fn session_file_build_works() {
-        let path = PathBuf::new();
-        let contents = get_template(&DateTime::now().to_formatted_pretty());
-        let file = SessionFile::build(&path, &contents).unwrap();
-        assert_eq!(&file.contents, &contents);
-    }
-
-    #[test]
-    fn session_file_get_heading_with_contents_works() {
-        let dt = &DateTime::now();
-        let contents = get_template(&dt.to_formatted_pretty());
-        let heading_contents = SessionFile::get_heading_with_contents(MARKS_HEADING, &contents);
-        assert_eq!(
-            heading_contents,
-            format!(
-                "\
-                {MARKS_HEADING}\n\
-                \n\
-                {MARK_HEADING_PREFIX}{}\
-                ",
-                dt.to_formatted_pretty()
-            )
-        );
-    }
-
-    #[test]
-    fn session_file_get_heading_level_works() {
-        assert_eq!(SessionFile::get_heading_level("# Heading"), 1);
-        assert_eq!(SessionFile::get_heading_level("## Heading"), 2);
-        assert_eq!(SessionFile::get_heading_level("##### Heading"), 5);
-    }
-
-    #[test]
-    fn session_new_works() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("."),
-        };
-        let dt = DateTime::now();
-        let mark = Mark::new(&dt.date);
-        let session = Session {
-            path: config
-                .sessions_path
-                .join(format!("{}.md", dt.to_formatted())),
-            marks: vec![mark],
-        };
-        assert_eq!(Session::new(&config, &DateTime::now()), session);
-    }
-
-    #[test]
-    fn session_start_works() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        let date = test_utils::now_plus_secs(30);
-        session.marks.push(Mark::new(&date));
-        assert_eq!(session.start(), session.marks.first().unwrap().date);
-    }
-
-    #[test]
-    fn session_end_works() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        assert_eq!(session.end(), session.marks.last().unwrap().date);
-        session.stop(&DateTime::now()).unwrap();
-        session.marks.last_mut().unwrap().date = test_utils::now_plus_secs(30);
-        assert_eq!(session.end(), session.marks.last().unwrap().date);
-    }
-
-    #[test]
-    fn session_is_active_works() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        assert!(session.is_active());
-        assert!(!session.marks.last().unwrap().has_label(&Label::End));
-        session.stop(&DateTime::now()).unwrap();
-        assert!(!session.is_active());
-        assert!(session.marks.last().unwrap().has_label(&Label::End));
-    }
-
-    // It ignores `mark_first` and counts to current time, so `mark_second` is the final time.
-    #[test]
-    fn session_get_time_ignores_marks_if_they_have_label_skip() {
-        let mut mark_first = Mark::new(&test_utils::now_plus_secs(-3 * 60 * 60));
-        mark_first.add_label(&Label::Skip);
-        let mark_second = Mark::new(&test_utils::now_plus_secs(-54 * 60 - 10)); // 54m 10s
-        let mark_third = Mark::new(&test_utils::now_plus_secs(-10 * 60));
-        let session = Session {
-            path: PathBuf::from("sessions"),
-            marks: vec![mark_first, mark_second, mark_third],
-        };
-        assert_eq!(session.get_time(), (54 * 60 + 10) * 1000);
-    }
-
-    #[test]
-    fn session_get_time_ignores_current_time_if_last_mark_has_label_skip() {
-        let mark_first = Mark::new(&test_utils::now_plus_secs(-3 * 60 * 60));
-        let mut mark_second = Mark::new(&test_utils::now_plus_secs(-1 * 60 * 60 - 33 * 60 - 20)); // 1h 33m 20s
-        mark_second.add_label(&Label::Skip);
-        let session = Session {
-            path: PathBuf::from("sessions"),
-            marks: vec![mark_first, mark_second],
-        };
-        assert_eq!(session.get_time(), (1 * 60 * 60 + 26 * 60 + 40) * 1000);
-    }
-
-    #[test]
-    fn session_stop_works() {
-        let config = Config {
-            action: Action::Stop {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        let mut clone = session.clone();
-        session.stop(&DateTime::now()).unwrap();
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        mark.add_label(&Label::End);
-        clone.marks.push(mark);
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn cannot_stop_when_session_ended() {
-        let config = Config {
-            action: Action::Stop {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        session.stop(&DateTime::now()).unwrap();
-        let clone = session.clone();
-        assert!(session.stop(&DateTime::now()).is_err());
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_mark_works() {
-        let dt = DateTime::now();
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        let mark = Mark::new(&dt.date);
-        session.mark(&DateTime::now()).unwrap();
-        assert_eq!(session.marks.len(), 2);
-        assert_eq!(session.marks[1], mark);
-    }
-
-    #[test]
-    fn session_mark_preserves_integrity_of_previous_content() {
-        let dt = DateTime::now();
-        let mark_first = Mark::new(&dt.date.with_hour(14).unwrap());
-        let mark_second = Mark::new(&mark_first.date.with_minute(47).unwrap());
-        let mut session = Session {
-            path: PathBuf::from(format!("./sessions/{}.md", dt.to_formatted())),
-            marks: vec![mark_first, mark_second],
-        };
-        let mut clone = session.clone();
-        session.mark(&DateTime::now()).unwrap();
-        assert_eq!(session.marks.len(), 3);
-        clone.marks.push(session.marks[2].clone());
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn cannot_mark_when_session_ended() {
-        let config = Config {
-            action: Action::Mark {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        session.stop(&DateTime::now()).unwrap();
-        let clone = session.clone();
-        assert!(session.mark(&DateTime::now()).is_err());
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_remark_works() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        session.mark(&DateTime::now()).unwrap();
-        let clone = session.clone();
-        session.marks.last_mut().unwrap().date = test_utils::now_plus_secs(30);
-        assert_ne!(session, clone);
-        session.remark(&DateTime::now());
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_label_works() {
-        let config = Config {
-            action: Action::Label { label: Label::Skip },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        // To have at least 2 marks.
-        session.mark(&DateTime::now()).unwrap();
-        let mut clone = session.clone();
-        session.label(&Label::Skip);
-        clone.marks.last_mut().unwrap().add_label(&Label::Skip);
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_unlabel_works() {
-        let config = Config {
-            action: Action::Label { label: Label::Skip },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        // To have at least 2 marks.
-        session.mark(&DateTime::now()).unwrap();
-        let clone = session.clone();
-        session.label(&Label::Skip);
-        session.unlabel(&Label::Skip);
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_write_works() {
-        let config = Config {
-            action: Action::Write {
-                text: String::new(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        // To have at least 2 marks.
-        session.mark(&DateTime::now()).unwrap();
-        let mut clone = session.clone();
-        session.write("hello").unwrap();
-        clone.marks.iter_mut().last().unwrap().write("hello");
-        assert_eq!(session, clone);
-    }
-
-    #[test]
-    fn session_write_errors_when_there_is_content() {
-        let config = Config {
-            action: Action::Start {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let mut session = Session::new(&config, &DateTime::now());
-        session.write("Some content.").unwrap();
-        assert!(session.write("Some other content.").is_err());
-    }
-
-    #[test]
-    fn session_from_file_works() {
-        let DateTime { date, .. } = DateTime::now();
-        let mark_first_dt = DateTime {
-            date: date.with_hour(5).unwrap(),
-            // FIX: same time for both marks breaks reading of them - doesn't read labels and keeps
-            // the whole content as is in the file
-            // date: date.with_hour(5).unwrap().with_minute(23).unwrap(),
-            // formatted: DateTime::format(&date.with_hour(5).unwrap().with_minute(23).unwrap()),
-        };
-        let mark_first = Mark::new(&mark_first_dt.date);
-        let mark_second_dt = DateTime {
-            date: mark_first.date.with_minute(23).unwrap(),
-        };
-        let mut mark_second = Mark::new(&mark_second_dt.date);
-        mark_second.add_label(&Label::End);
-
-        let contents = format!(
-            "\
-                {SESSION_HEADING_PREFIX}{SESSION_TITLE}\n\
-                \n\
-                {MARKS_HEADING}\n\
-                \n\
-                {MARK_HEADING_PREFIX}{}\n\
-                \n\
-                {MARK_HEADING_PREFIX}{}\n\
-                \n\
-                {LABEL_END}\n\
-                ",
-            mark_first_dt.to_formatted_pretty(),
-            mark_second_dt.to_formatted_pretty(),
-        );
-        let file = SessionFile::build(&PathBuf::new(), &contents).unwrap();
-        let session = Session {
-            path: file.path.clone(),
-            marks: vec![mark_first, mark_second],
-        };
-
-        assert_eq!(Session::from_file(&file).unwrap(), session);
-    }
-
-    #[test]
-    fn session_from_file_fails_when_there_is_no_mark() {
-        let contents = format!(
-            "\
-                {SESSION_HEADING_PREFIX}{SESSION_TITLE}\n\
-                \n\
-                {MARKS_HEADING}\n\
-                ",
-        );
-        let file = SessionFile::build(&PathBuf::from("sessions"), &contents).unwrap();
-        assert!(Session::from_file(&file).is_err());
-    }
-
-    #[test]
-    fn session_to_file_works() -> Result<(), Box<dyn Error>> {
-        let dt = DateTime::now();
-        let mark_first_dt = DateTime {
-            date: dt.date.with_hour(5).unwrap(),
-        };
-        let mark_first = Mark::new(&mark_first_dt.date);
-        let mark_second_dt = DateTime {
-            date: mark_first_dt.date.with_minute(44).unwrap(),
-        };
-        let mark_second = Mark {
-            date: mark_second_dt.date,
-            labels: HashSet::new(),
-            contents: String::from("I am the second mark!\nHi!\n"),
-        };
-        let config = Config {
-            action: Action::Mark {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let session = Session {
-            path: config
-                .sessions_path
-                .join(format!("{}.md", mark_first_dt.to_formatted())),
-            marks: vec![mark_first, mark_second],
-        };
-        let file = SessionFile::build(
-            &session.path,
-            &format!(
-                "\
-                    {SESSION_HEADING_PREFIX}{SESSION_TITLE}\n\
-                    \n\
-                    {MARKS_HEADING}\n\
-                    \n\
-                    {MARK_HEADING_PREFIX}{}\n\
-                    \n\
-                    {MARK_HEADING_PREFIX}{}\n\
-                    \n\
-                    I am the second mark!\n\
-                    Hi!\n\
-                    ",
-                mark_first_dt.to_formatted_pretty(),
-                mark_second_dt.to_formatted_pretty()
-            ),
-        )?;
-
-        assert_eq!(session.to_file()?, file);
-        Ok(())
-    }
-
-    #[test]
-    fn session_to_file_and_from_file() {
-        let dt = DateTime::now();
-        let mark_first = Mark {
-            date: dt.date.with_hour(5).unwrap().with_minute(54).unwrap(),
-            labels: HashSet::new(),
-            contents: String::from("feat/some-branch\n\nDid a few things"),
-        };
-        let mark_second = Mark {
-            date: dt.date.with_hour(6).unwrap().with_minute(13).unwrap(),
-            labels: HashSet::new(),
-            contents: String::from("feat/new-feature"),
-        };
-        let config = Config {
-            action: Action::Mark {
-                date: DateTime::now(),
-            },
-            sessions_path: PathBuf::from("sessions"),
-        };
-        let session = Session {
-            path: config
-                .sessions_path
-                .join(&format!("{}.md", DateTime::format(&mark_first.date))),
-            marks: vec![mark_first, mark_second],
-        };
-        let file = session.to_file().unwrap();
-        assert_eq!(Session::from_file(&file).unwrap(), session);
-    }
-
-    #[test]
-    fn mark_new_works() {
-        let dt = DateTime::now();
-        let mark = Mark::new(&dt.date);
-        assert_eq!(Mark::new(&dt.date), mark);
-    }
-
-    #[test]
-    fn mark_add_label_works() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        mark.add_label(&Label::End);
-        let expected = Mark {
-            date: dt.date,
-            labels: HashSet::from_iter([Label::End]),
-            contents: String::new(),
-        };
-        assert_eq!(mark, expected);
-    }
-
-    #[test]
-    fn mark_add_label_does_not_add_duplicate_label() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        mark.add_label(&Label::Skip);
-        let clone = mark.clone();
-        mark.add_label(&Label::Skip);
-        assert_eq!(mark, clone);
-    }
-
-    #[test]
-    fn mark_remove_label_works() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        let clone = mark.clone();
-        mark.add_label(&Label::Skip);
-        mark.remove_label(&Label::Skip);
-        assert_eq!(mark, clone);
-        mark.remove_label(&Label::Skip);
-        assert_eq!(mark, clone);
-    }
-
-    #[test]
-    fn mark_has_label_works() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        mark.add_label(&Label::End);
-        assert!(mark.has_label(&Label::End));
-    }
-
-    #[test]
-    fn mark_write_works() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        let mut clone = mark.clone();
-        clone.contents = String::from("This is some content.");
-        mark.write("This is some content.");
-        assert_eq!(mark, clone);
-    }
-
-    #[test]
-    fn mark_write_overwrites_previous_content() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        let mut clone = mark.clone();
-        clone.contents = String::from("This is some content");
-        mark.write("This is some content");
-        assert_eq!(mark, clone);
-        clone.contents = String::from("This is overwritten content");
-        mark.write("This is overwritten content");
-        assert_eq!(mark, clone);
-    }
-
-    #[test]
-    fn mark_erase_works() {
-        let dt = DateTime::now();
-        let mut mark = Mark::new(&dt.date);
-        let clone = mark.clone();
-        mark.write("This is content.");
-        mark.erase();
-        assert_eq!(mark, clone);
-    }
-
-    #[test]
-    fn mark_from_string_works() {
-        let dt = DateTime::now();
-        let contents = format!(
-            "\
-                {MARK_HEADING_PREFIX}{}\n\
-                \n\
-                {LABEL_END}\n\
-                \n\
-                This is some content.\n\
-                ",
-            dt.to_formatted_pretty()
-        );
-        let mark = Mark {
-            date: dt.date,
-            labels: HashSet::from_iter([Label::End]),
-            contents: String::from("This is some content."),
-        };
-        assert_eq!(Mark::from_string(&contents).unwrap(), mark);
-    }
-
-    #[test]
-    fn mark_to_string_works() {
-        let dt = DateTime::now();
-        let mark = Mark {
-            date: dt.date,
-            labels: HashSet::from_iter([Label::End]),
-            contents: String::from("This is a content of a mark.\nHow are you?\n"),
-        };
-        let output = format!(
-            "\
-                {MARK_HEADING_PREFIX}{}\n\
-                \n\
-                {LABEL_END}\n\
-                \n\
-                This is a content of a mark.\n\
-                How are you?\
-                ",
-            dt.to_formatted_pretty()
-        );
-        assert_eq!(mark.to_string(), output);
-    }
-
-    #[test]
-    fn label_from_args_works() -> Result<(), Box<dyn Error>> {
-        assert!(Label::from_args(&[]).is_err());
-        assert!(Label::from_args(&[String::from("hello")]).is_err());
-
-        assert_eq!(Label::from_args(&[String::from("end")])?, Label::End);
-        assert!(Label::from_args(&[String::from("end"), String::from("hello")]).is_err());
-
-        assert_eq!(Label::from_args(&[String::from("skip")])?, Label::Skip);
-        assert!(Label::from_args(&[String::from("skip"), String::from("hello")]).is_err());
-
-        assert!(Label::from_args(&[String::from("tag")]).is_err());
-        assert_eq!(
-            Label::from_args(&[String::from("tag"), String::from("rust")])?,
-            Label::Tag {
-                text: String::from("rust")
-            }
-        );
-        assert!(Label::from_args(&[
-            String::from("tag"),
-            String::from("rust"),
-            String::from("hello")
-        ])
-        .is_err());
-
-        Ok(())
-    }
-
-    #[test]
-    fn label_from_string_works() {
-        assert_eq!(Label::from_string(LABEL_END).unwrap(), Label::End);
-
-        assert_eq!(Label::from_string(LABEL_SKIP).unwrap(), Label::Skip);
-
-        assert!(Label::from_string(&format!(
-            "{LABEL_TAG} {LABEL_TAG_SURROUND}{LABEL_TAG_SURROUND}"
-        ))
-        .is_err());
-        assert!(Label::from_string(&format!("{LABEL_TAG} {LABEL_TAG_SURROUND}rust")).is_err());
-        assert!(Label::from_string(&format!("rust{LABEL_TAG_SURROUND}")).is_err());
-        assert_eq!(
-            Label::from_string(&format!(
-                "{LABEL_TAG} {LABEL_TAG_SURROUND}rust{LABEL_TAG_SURROUND}"
-            ))
-            .unwrap(),
-            Label::Tag {
-                text: String::from("rust")
-            }
-        );
-
-        assert!(Label::from_string("some string").is_err())
-    }
-
-    #[test]
-    fn label_to_string_works() {
-        assert_eq!(Label::End.to_string(), LABEL_END);
-
-        assert_eq!(Label::Skip.to_string(), LABEL_SKIP);
-
-        assert_eq!(
-            Label::Tag {
-                text: String::from("rust"),
-            }
-            .to_string(),
-            format!("{LABEL_TAG} {LABEL_TAG_SURROUND}rust{LABEL_TAG_SURROUND}")
-        );
-    }
-
-    #[test]
-    fn label_to_string_and_from_string() {
-        let label = Label::End;
-        let as_string = label.to_string();
-        assert_eq!(Label::from_string(&as_string).unwrap(), label);
     }
 }
